[Feat] String 입력받아서 가격형태로 변환시켜주는 convertPrice함수 String extension으로 분리 - 원래는 사용되는 뷰마다 적어줬는데, 이게 좀 더 깔끔하고 재사용성이 좋아보여서!
